✅ (web) add collectCoverageFrom Jest configuration
An array of glob patterns indicating a set of files for which coverage information should be collected. If a file matches the specified glob pattern, coverage information will be collected for it even if no tests exist for this file and it's never required in the test suite.
Note: This option requires collectCoverage to be set to true or Jest to be invoked with --coverage.